# CheChe/login-service/src/main/java/com/cheche/login/dto/CredentialsRequest.java
package com.cheche.login.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Pattern;
import jakarta.validation.constraints.Size;

public record CredentialsRequest(
        @NotBlank
        @Size(min = 4, max = 50)
        @Pattern(regexp = "^[a-zA-Z0-9._-]+$", message = "아이디는 영문, 숫자, 점, 밑줄, 하이픈만 사용할 수 있습니다.")
        String username,

        @NotBlank
        @Size(min = 8, max = 72, message = "비밀번호는 8자 이상 72자 이하여야 합니다.")
        String password
) {}