# CheChe/login-service/src/main/java/com/cheche/login/dto/UserLoginResponse.java
package com.cheche.login.dto;

public record UserLoginResponse(
        Long userId,
        String username,
        String accountType,
        String tokenType,
        String accessToken,
        long expiresInSeconds,
        String message
) {}